perf: dynamically choose throtted or debounced
